requirements.txt
pymongo[srv]

langchain
langchain_groq
langchain_community
langchain_core
langsmith
python-dotenv
pymongo
flask
flask-cors
langchain_openai
google-generativeai


